BP Nouveau: generate the Activity container before adding new entries
If the Activity container is missing, make sure to generate it before prepending new activities. If the container is available the activity will only be inserted if the activity stream can welcome it. For instance, an activity posted in a Private group will be prepended to the list only if the "My Groups" scope is active.

Fixes #7774

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12038 cdf35c40-ae34-48e0-9cc9-0c9da1808c22