BP Nouveau: display the Group Invites UI navigation vertically
On desktop screens, the Group Invites UI navigation is now displayed vertically at the left of the friends/members list or the send invites form. This new position makes it easier to navigate between the different screens and improve the visibility of the send invites navigation item.

Props mercime, boonebgorges.

Fixes #7783
Fixes #7787

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12040 cdf35c40-ae34-48e0-9cc9-0c9da1808c22