BP Nouveau: restrict the allowed HTML tags for Messages.
BP Nouveau is using the WP Editor to improve the user experience of the Messages component when people are writing a new message or replying to a thread. So far allowed HTML tags for the messages content *when BP Nouveau is the active template pack* were the same than WordPress Posts. When BP Nouveau is not the active template pack, the allowed tags are those of the WordPress global .

This commit is introducing a new function  that is used to filter the allowed HTML tags for activity content and message content. It extends the  with new tags (img, span, ul, ol & li) and leaves specific functions to the 2 components to include filters so that it is possible to restrict/extend allowed HTML tags for both content types or one of them. It also restricts the WP Editor available buttons when used into the BP Nouveau Messages UI so that they are consistent with the allowed HTML tags.

Props DJPaul

Fixes #7795

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12067 cdf35c40-ae34-48e0-9cc9-0c9da1808c22