BP Nouveau - Customizer: improve Group activity stream preview.
Single Group Activity stream was missing a class when the Group front page Customizer setting was disabled. In this particular case, the Group home page is the Group Activity stream page. The class was not added because the current action is not "activity" but "home". Using the {{{bp_is_group_activity()}}} conditional tag solves the issue.

Props dunhakdis

Fixes #7813

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12075 cdf35c40-ae34-48e0-9cc9-0c9da1808c22