BP Nouveau Customizer: Improve Primary Nav Item reordoring control.
Remove the restriction about the top/first primary nav item to let the site owner control the order of each items. It now allows to be consistent with the {{{BP_DEFAULT_COMPONENT}}} if the custom front is not used by moving this default component nav item at the first position.

Props sbrajesh, hnla

See #7871 (Trunk)

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12137 cdf35c40-ae34-48e0-9cc9-0c9da1808c22