Nouveau: fix filtering Member Directory when specified by user type.
This is a problem we had with BP-Legacy's JS, and as it seems this part was
copied into Nouveau, we are implementing the same fix here. This involves not
using Javascript filtering for the Member Directory filters when a user type
is set.

A better fix would be to rewrite the Javascript to solve the underlying issue.

Fixes #7903

Props imath

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12224 cdf35c40-ae34-48e0-9cc9-0c9da1808c22