Nav: Dynamically load component screen file when changing nav default.
BP 3.0.0 moved most screen controllers into files that are loaded only when
viewing the component. This introduced a regression into
`bp_core_new_nav_default()`, since the `screen_function` would not be available
(ie `! is_callable()`) at the time that the nav item is modified.

We resolve this by loading the component's screen functions in
`bp_core_new_nav_default()` when we detect that they're not present. To work
around cases where component slugs don't match the official component ID,
which corresponds to the file path, we introduce a new 'component_id' parameter
to be used when registering a new nav item.

Props imath, r-a-y.
Fixes #7935.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12288 cdf35c40-ae34-48e0-9cc9-0c9da1808c22