Activity: Fix parameters for instance of `bp_activity_custom_update` filter.
This filter, introduced in [2552], is intended to allow developers to specify a
custom activity ID after posting to the "what's new" box. As originally
written, the filter didn't pass the right parameters to be properly used.
The instances of the filter that appear in bp-templates files were fixed in
[9160]; see #6021. The current changeset fixes the remaining instance.

Fixes #8054.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12355 cdf35c40-ae34-48e0-9cc9-0c9da1808c22