Improve i18n of Groups activity updates action string
Dynamically generated activity action strings were introduced in BuddyPress 2.0 to ensure these strings are always up to date and multilingual-friendly (see #3856).

It appeared although the {{{activity_update}}} type enjoys this feature when activities are shared by users on their profiles, it was not the case for activity updates posted within Groups.

Thanks to this commit, BuddyPress is now taking it in charge by:
- Registering a new group activity action to reference a specific formatting callback function for activities posted within Groups.
- Making sure this new activity action type does not interfere with the one of regular activity updates in dropdown filters or into the Activity Administration screens.

BTW happy 4th WPTranslationDay to everyone :)

Fixes #8089

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12395 cdf35c40-ae34-48e0-9cc9-0c9da1808c22