BP Settings: Introduce a new BP friendly name key into data exporters
The name of this key is {{{exporter_bp_friendly_name}}}. It is added to BP & WP exporter arrays to adapt the existing {{{exporter_friendly_name}}} for a non Administrator role context. If this new key is provided it will be used to output the friendly name of the exported items, otherwise it will use the existing {{{exporter_friendly_name}}} key.

For example, into the items list of the {{{site.url/username/settings/data}}} page of the logged in user, "WordPress Comments" will be replaced by "Comments" and so on for every built-in WordPress and BuddyPress items.

Props sbrajesh, r-a-y, boonebgorges

Fixes #8013

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12416 cdf35c40-ae34-48e0-9cc9-0c9da1808c22