Cast `func_get_args()` to variable early in class methods.
Since PHP 7.0, calling `func_get_args()` returns the current value of
the parameter, which may have been modified since the beginning of the
method. By storing the original function arguments in a variable at the
beginning of the method, we ensure that the value operated on later in
the method accurately reflects what was originally passed.

In most cases addressed by this changeset, there was no real-world
possibility that the arguments could be changed; we make this change
mostly as a best practice, and to appease the PHP compatibility linter.

Fixes #8125.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12426 cdf35c40-ae34-48e0-9cc9-0c9da1808c22