Add node modules to have tools to compile & bundle modern JavaScript
1. Add Parcel and Babel
Parcel is a web application bundler used to watch and build BP Blocks and BP Block components. It is associated with Babel, a toolchain that is mainly used to convert ECMAScript 2015+ code into a backwards compatible version of JavaScript in current and older browsers or environments.

2. Add the default Babel preset for WordPress development.

3. Remove our development dependency to `grunt-wp-i18n` as we will use from now on WP CLI to generate a pot file that is also taking in account JavaScript i18n strings.

4. Add the needed Parcel commands to prepare BP Blocks compiling tasks.

5. Update Grunt.js to latest stable.

NB:  contributors, please use `npm install` to update your node modules locally.

See #8048
See #8154


git-svn-id: https://buddypress.svn.wordpress.org/trunk@12573 cdf35c40-ae34-48e0-9cc9-0c9da1808c22