Add an extra Groups menu to site's admin if BP is network activated
On multisite configs, as it's not possible to edit tags into the network admin, BuddyPress is generating the sub menu item to manage Group Types into the main site's admin. In this case, and this case only, we also need to generate a parent menu item to redirect the Administrator to the Network Groups admin page.

Fixes #8372


git-svn-id: https://buddypress.svn.wordpress.org/trunk@12777 cdf35c40-ae34-48e0-9cc9-0c9da1808c22