xProfile: introduce the Checkbox Acceptance field type
This new field type helps community site administrators get members agreement about specific rules. It can be terms of service, code of conduct, game rules, etc.

When adding such a field, the xProfile field Admin screen will contain a specific metabox to select the WordPress page to use as the linked content needing members agreement.

This field can be required or not, but cannot be updated once members agreed to the WordPress page it relies to.

Props mahdiar, vapvarun

Fixes #4747



git-svn-id: https://buddypress.svn.wordpress.org/trunk@12882 cdf35c40-ae34-48e0-9cc9-0c9da1808c22