Nouveau: only handle Messages supported views with the Backbone router
When an unsupported view is generated by a plugin (eg: adding a new Messages Component sub nav item), it is skipped by the Backbone router and the link attached to it is loaded the traditional way on click. If the current screen has been added by a plugin, the sub nav click handler also leaves the event loads the page traditionally.

Props Venutius

Fixes #8430



git-svn-id: https://buddypress.svn.wordpress.org/trunk@12941 cdf35c40-ae34-48e0-9cc9-0c9da1808c22