Nouveau: only unregister the 'bp/primary-nav' block if it's registered
Nouveau unregisters this block when an admin edits a post type as it doesn't make sense to use this block into the Post Block Editor context. To prevent a `_doing_it_wrong()` from WordPress, it's probably safest to check if the block exists before unregistering it.

This commits also makes sure the user is using WP > 5.0.

Fixes #8528



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13016 cdf35c40-ae34-48e0-9cc9-0c9da1808c22