Settings: Make sure changing pwd from the General Screen encrypts it
Using a WP_User object to set the password to update was misleading WordPress `wp_update_user()` function. The new password was interpreted as the current password and was not encrypted.

Passing a regular array to the `wp_update_user()` function avoids the risk of messing with WP User object cache.

Props mandro

See #8539 (branch 9.0)



git-svn-id: https://buddypress.svn.wordpress.org/branches/9.0@13031 cdf35c40-ae34-48e0-9cc9-0c9da1808c22