Admin: improve information about WP pages being used as BP directories
We are now using BuddyPress specific post states to inform a WordPress page is used as the BP Component's directory page. This information is displayed into the WP Admin Pages edit screen, for menu items linking to a BP Directory into the WP Admin Menu screen as well as into the Customizer Menu section.

Fixes #8470


git-svn-id: https://buddypress.svn.wordpress.org/trunk@13123 cdf35c40-ae34-48e0-9cc9-0c9da1808c22