Make sure the `new_avatar` activity can use the avatar crop results
Since [13175] the `BP_Attachment_Avatar::crop()` method include the timestamp the avatar was generated on into its returned array. This array is now transported into the action the function creating a `new_avatar` activity hooks to. This function (`bp_members_new_avatar_activity()`) now accepts 3 more parameters included the transported crop results. This makes it possible to use the timestamp when the avatar was generated on as the `new_avatar` activity recorded time. Doing so, we don't need to add an extra activity meta to store the name of the avatar file to be sure to display the avatar the user had at the time the activity was created.

Props vapvarun, oztaser

See #8581



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13177 cdf35c40-ae34-48e0-9cc9-0c9da1808c22