Improve group members count query performance
[13103] introduced a change in 10.0.0 that can be very time consuming when a group has a lot of members.

To keep the main improvements of the referenced commit (only refreshing group members count when a user joins or leaves a group) but optimize queries performance, we are introducing a new way to count group members in the `BP_Group_Member_Query`.

We are also introducing a way to defer group members count when adding a batch of members to a group. Using `bp_groups_defer_group_members_count()` avoids to refresh the count each time a member of this batch is added. For more information about how to use this function, you can have a look at how BuddyPress is using it into `src/bp-groups/bp-groups-admin.php`.

Props dd32, espellcaste

See #8688 (trunk)



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13280 cdf35c40-ae34-48e0-9cc9-0c9da1808c22