Make sure `BP_Component::parse_query()` always run
Only running it when BP Rewrites are on is not the right way to go as some plugins might have been using the WP Rewrites API for a while and might have extended this method from their component's class.

Instead of only running this method if BP Rewrites are on, we're adding a URL parser check inside each component having a directory to figure out whether it's needed to set BP URI globals.

See #4954
Fixes #8908
Closes https://github.com/buddypress/buddypress/pull/109



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13492 cdf35c40-ae34-48e0-9cc9-0c9da1808c22