Add 2 new filterable functions to get register/activate root slugs
These 2 functions are used to set rewrite rules and permastructs for registration and activation pages.

Using the provided filters it is now possible to prepend a `custom-slug/` to these generated rewrite rules and permastructs. This was the only missing thing to be able to prepend a `custom-slug/` to all BP directory/specific page root slugs.

Props espellcaste, slaFFik, dcavins.

See #9063
Closes https://github.com/buddypress/buddypress/pull/364



git-svn-id: https://buddypress.svn.wordpress.org/trunk@14011 cdf35c40-ae34-48e0-9cc9-0c9da1808c22