BP Legacy: load newest activities according to displayed context
Make sure the `bp_heartbeat` property is included into the sent Heartbeat data object & use it to transport the cookies BP Legacy uses to remember logged in user's navigation preferences.

The `bp_heartbeat` property is needed to be sure the current Heartbeat request is about BuddyPress and needs to be parsed using the BP Rewrite API since [13752]. This parsing job analyses the current URL to define the value of BP URI globals such as the current displayed single item.

Props dreampixel

Fixes #9248. (14.0 branch)

git-svn-id: https://buddypress.svn.wordpress.org/branches/14.0@14083 cdf35c40-ae34-48e0-9cc9-0c9da1808c22